Added Audio (Real)
Waiting on the voice lines and full music, as well as having possibly more sound effects. I got a bit tripped up by the fact that I couldn't put the sounds in the same scene as the levels, but it ended up being better that way anyways. Also ignore the branch I made yesterday, this is the one that actually works